# src/public/swagger/swagger.yaml
swagger: '2.0'
info:
  description: QuantSwing — Automated Intraday Options Trading Platform, Backend API
  version: 1.0.0
  title: QuantSwing Backend
basePath: /
schemes:
  - http
  - https
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT access token, sent as: `Bearer <token>`'
  Token:
    type: apiKey
    name: token
    in: header
    description: 'Raw JWT access token (hedged-style `token` header). Alternative to BearerAuth.'
paths:
...
  /trade-setups:
    post:
      operationId: createTradeSetup
      summary: Create a trade setup (name + legs)
      tags: [trade-setup]
      consumes: [application/json]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreateTradeSetupRequest'
      responses:
        201:
          description: Created setup with legs
          schema:
            $ref: '#/definitions/TradeSetup'
        400:
          description: Invalid legs (lot size / freeze quantity)
          schema:
            $ref: '#/definitions/Error'
    get:
      operationId: listTradeSetups
      summary: List trade setups
      tags: [trade-setup]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: status
          in: query
          required: false
          type: string
          enum: [ACTIVE, ARCHIVED]
      responses:
        200:
          description: Setups with legs and live-position counts
          schema:
            type: array
            items:
              $ref: '#/definitions/TradeSetup'
  /trade-setups/{id}:
    get:
      operationId: getTradeSetup
      summary: Get one trade setup
      tags: [trade-setup]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: The setup
          schema:
            $ref: '#/definitions/TradeSetup'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/Error'
    patch:
      operationId: updateTradeSetup
      summary: Update name/status/legs (legs only when no live position)
      tags: [trade-setup]
      consumes: [application/json]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateTradeSetupRequest'
      responses:
        200:
          description: Updated setup
          schema:
            $ref: '#/definitions/TradeSetup'
        409:
          description: Live position references this setup
          schema:
            $ref: '#/definitions/Error'
    delete:
      operationId: deleteTradeSetup
      summary: Delete a setup (only when never used)
      tags: [trade-setup]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: Deleted
          schema:
            $ref: '#/definitions/MessageResponse'
        409:
          description: Referenced by positions
          schema:
            $ref: '#/definitions/Error'
  /positions:
    post:
      operationId: createPosition
      summary: Take a position on a trade setup
      tags: [position]
      consumes: [application/json]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreatePositionRequest'
      responses:
        201:
          description: Position with legs
          schema:
            $ref: '#/definitions/Position'
        409:
          description: A live position already exists for this setup
          schema:
            $ref: '#/definitions/Error'
    get:
      operationId: listPositions
      summary: List positions (newest first)
      tags: [position]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: status
          in: query
          required: false
          type: string
          enum: [OPEN, MARKED_FOR_EXIT, CLOSED]
      responses:
        200:
          description: Positions with legs, orders and setup
          schema:
            type: array
            items:
              $ref: '#/definitions/Position'
  /positions/{id}:
    get:
      operationId: getPosition
      summary: Get one position
      tags: [position]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: The position
          schema:
            $ref: '#/definitions/Position'
        404:
          description: Not found
          schema:
            $ref: '#/definitions/Error'
    patch:
      operationId: updatePosition
      summary: Mark for exit or change the multiplier
      description: >
        Setting status MARKED_FOR_EXIT flips every leg's target quantity to 0 —
        the order dialog then generates the reverse (square-off) orders.
      tags: [position]
      consumes: [application/json]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdatePositionRequest'
      responses:
        200:
          description: Updated position
          schema:
            $ref: '#/definitions/Position'
        409:
          description: Position closed / invalid transition
          schema:
            $ref: '#/definitions/Error'
    delete:
      operationId: deletePosition
      summary: Delete a position and its orders (no live orders allowed)
      tags: [position]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        200:
          description: Deleted
          schema:
            $ref: '#/definitions/MessageResponse'
        409:
          description: Live orders present
          schema:
            $ref: '#/definitions/Error'
  /positions/{id}/orders:
    post:
      operationId: placePositionOrders
      summary: Place a batch of diff orders against the position's legs
      description: >
        The legacy place_order flow — the client computes the required orders
        from held vs target quantities (sync) or reverse orders (exit) and
        posts them together. Orders carrying an `id` modify that OPEN order.
        Placement is risk-sorted (exits/hedges first). Market hours only.
      tags: [position]
      consumes: [application/json]
      produces: [application/json]
      security:
        - BearerAuth: []
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/PlacePositionOrdersRequest'
      responses:
        201:
          description: Persisted orders in placement order
          schema:
            type: object
            properties:
              orders:
                type: array
                items:
                  $ref: '#/definitions/Order'
        400:
          description: Market closed / validation failure
          schema:
            $ref: '#/definitions/Error'
        409:
          description: Position closed or order no longer OPEN
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  Instrument:
    type: object
    title: Instrument
    description: One tradable/watchable contract from the filtered Angel One scrip master.
    properties:
      id:
        type: string
        description: '`${exchSeg}:${symbol}`'
        example: 'NFO:NIFTY29JAN2621000CE'
      token:
        type: string
        description: Angel One instrument token (used for the WS feed and orders)
      symbol:
        type: string
        example: NIFTY29JAN2621000CE
      name:
        type: string
        description: Underlying name
        example: NIFTY
      expiry:
        type: string
        format: date
        x-nullable: true
      strike:
        type: number
        example: 21000
      lotSize:
        type: integer
      tickSize:
        type: number
      instrumentType:
        type: string
        enum:
          - OPTIDX
          - AMXIDX
      exchSeg:
        type: string
        enum:
          - NFO
          - BFO
          - NSE
          - BSE
      freezeQty:
        type: integer
      lastPrice:
        type: number
        description: Last traded price, refreshed by the LTP stream
      volume:
        type: number
        description: Cumulative day volume, refreshed by the LTP stream
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
...
  MessageResponse:
    type: object
    title: MessageResponse
    properties:
      message:
        type: string
        example: Logged out
...
  TradeSetupLeg:
    type: object
    title: TradeSetupLeg
    properties:
      id:
        type: string
        format: uuid
      tradeSetupId:
        type: string
        format: uuid
      instrumentId:
        type: string
      instrument:
        $ref: '#/definitions/Instrument'
      action:
        type: string
        enum: [BUY, SELL]
      quantity:
        type: integer
        description: Absolute target quantity at multiplier 1 (lots x lotSize)
      entryPrice:
        type: number
        x-nullable: true
      exitPrice:
        type: number
        x-nullable: true
      createdAt:
        type: string
        format: date-time
  TradeSetup:
    type: object
    title: TradeSetup
    description: User-defined multi-leg trade template — the reconcile target for positions.
    properties:
      id:
        type: string
        format: uuid
      name:
        type: string
      status:
        type: string
        enum: [ACTIVE, ARCHIVED]
      legs:
        type: array
        items:
          $ref: '#/definitions/TradeSetupLeg'
      livePositions:
        type: integer
        description: Count of non-CLOSED positions referencing this setup (list endpoint only)
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  TradeSetupLegInput:
    type: object
    title: TradeSetupLegInput
    required: [instrumentId, action, quantity]
    properties:
      instrumentId:
        type: string
      action:
        type: string
        enum: [BUY, SELL]
      quantity:
        type: integer
        description: Must be a lot-size multiple
      entryPrice:
        type: number
      exitPrice:
        type: number
  CreateTradeSetupRequest:
    type: object
    title: CreateTradeSetupRequest
    required: [name, legs]
    properties:
      name:
        type: string
      legs:
        type: array
        items:
          $ref: '#/definitions/TradeSetupLegInput'
  UpdateTradeSetupRequest:
    type: object
    title: UpdateTradeSetupRequest
    properties:
      name:
        type: string
      status:
        type: string
        enum: [ACTIVE, ARCHIVED]
      legs:
        type: array
        description: Full replacement; only allowed with no live position
        items:
          $ref: '#/definitions/TradeSetupLegInput'
  PositionLeg:
    type: object
    title: PositionLeg
    description: >
      One leg of a position. Carries no stored quantity — held quantity is
      derived from its COMPLETED orders.
    properties:
      id:
        type: string
        format: uuid
      positionId:
        type: string
        format: uuid
      tradeSetupLegId:
        type: string
        format: uuid
      tradeSetupLeg:
        $ref: '#/definitions/TradeSetupLeg'
      instrumentId:
        type: string
      instrument:
        $ref: '#/definitions/Instrument'
      action:
        type: string
        enum: [BUY, SELL]
      orders:
        type: array
        items:
          $ref: '#/definitions/Order'
      createdAt:
        type: string
        format: date-time
  Position:
    type: object
    title: Position
    description: >
      A live instance of a TradeSetup. `position-update` socket events carry
      this shape after every reconcile.
    properties:
      id:
        type: string
        format: uuid
      userId:
        type: string
        format: uuid
      tradeSetupId:
        type: string
        format: uuid
      tradeSetup:
        $ref: '#/definitions/TradeSetup'
      broker:
        type: string
        enum: [PAPER]
      status:
        type: string
        enum: [OPEN, MARKED_FOR_EXIT, CLOSED]
      userMultiplier:
        type: integer
      readableStatus:
        type: string
        x-nullable: true
        enum: [SYNCED, NOT_IN_SYNC, IN_PROGRESS, EXITED]
      pnl:
        type: number
        description: Realized P&L (MTM poller output)
      calcUnrealisedPnl:
        type: number
        description: Unrealized MTM on remaining quantity
      totalCharges:
        type: number
      legs:
        type: array
        items:
          $ref: '#/definitions/PositionLeg'
      exitedAt:
        type: string
        format: date-time
        x-nullable: true
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  CreatePositionRequest:
    type: object
    title: CreatePositionRequest
    required: [tradeSetupId]
    properties:
      tradeSetupId:
        type: string
      userMultiplier:
        type: integer
        default: 1
  UpdatePositionRequest:
    type: object
    title: UpdatePositionRequest
    description: At least one of status or userMultiplier is required.
    properties:
      status:
        type: string
        enum: [MARKED_FOR_EXIT]
      userMultiplier:
        type: integer
  PositionOrderInput:
    type: object
    title: PositionOrderInput
    required: [positionLegId, transactionType, orderType, quantity]
    properties:
      id:
        type: string
        description: Present — modify this OPEN order instead of creating one
      positionLegId:
        type: string
      transactionType:
        type: string
        enum: [BUY, SELL]
      orderType:
        type: string
        enum: [MARKET, LIMIT]
      quantity:
        type: integer
      price:
        type: number
        description: Required for LIMIT
      productType:
        type: string
        enum: [INTRADAY, NORMAL]
        default: INTRADAY
  PlacePositionOrdersRequest:
    type: object
    title: PlacePositionOrdersRequest
    required: [orders]
    properties:
      orders:
        type: array
        items:
          $ref: '#/definitions/PositionOrderInput'
...
  Order:
    type: object
    title: Order
    description: >
      A single order routed to a broker. Status flow: PENDING (queued) -> OPEN
      (accepted) -> COMPLETED/CANCELLED/REJECTED. `order-update` socket events
      fire on every transition.
    properties:
      id:
        type: string
        format: uuid
      userId:
        type: string
        format: uuid
      instrumentId:
        type: string
        example: 'NFO:NIFTY29JAN2621000CE'
      instrument:
        $ref: '#/definitions/Instrument'
      broker:
        type: string
        enum:
          - PAPER
      status:
        type: string
        enum:
          - PENDING
          - OPEN
          - PARTIAL
          - COMPLETED
          - CANCELLED
          - REJECTED
      transactionType:
        type: string
        enum:
          - BUY
          - SELL
      orderType:
        type: string
        description: MARKET orders are resolved to LIMIT at placement time
        enum:
          - MARKET
          - LIMIT
      productType:
        type: string
        enum:
          - INTRADAY
          - NORMAL
      quantity:
        type: integer
      price:
        type: number
        description: Limit price
      brokerOrderId:
        type: string
        x-nullable: true
        description: Broker-side order id; null until accepted
        example: PAPER-8fa85a5d-a2b2-4b1f-899e-5014436056ef
      filledQuantity:
        type: integer
      averageExecutionPrice:
        type: number
        x-nullable: true
      rejectReason:
        type: string
        x-nullable: true
      placedAt:
        type: string
        format: date-time
        x-nullable: true
      executedAt:
        type: string
        format: date-time
        x-nullable: true
      cancelledAt:
        type: string
        format: date-time
        x-nullable: true
      brokerLogs:
        type: array
        description: Present only on GET /orders/{id} (up to 20 recent)
        items:
          $ref: '#/definitions/BrokerLog'
      positionId:
        type: string
        x-nullable: true
        description: OMS linkage; null for quick-ticket orders
      positionLegId:
        type: string
        x-nullable: true
      stt:
        type: number
        x-nullable: true
      stampDuty:
        type: number
        x-nullable: true
      exchangeCharges:
        type: number
        x-nullable: true
      sebiCharges:
        type: number
        x-nullable: true
      brokerage:
        type: number
        x-nullable: true
      gst:
        type: number
        x-nullable: true
      totalCharges:
        type: number
        x-nullable: true
        description: Sum of all charges, stamped on COMPLETED fills
      chase:
        type: boolean
        description: Whether the price chase re-quotes this order while OPEN
      chaseAttempts:
        type: integer
      chaseFailures:
        type: integer
      remarks:
        type: string
        x-nullable: true
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  BrokerLog:
    type: object
    title: BrokerLog
    description: Audit trail of one broker API call (real or simulated).
    properties:
      id:
        type: string
        format: uuid
      broker:
        type: string
        enum:
          - PAPER
      method:
        type: string
        example: POST
      url:
        type: string
        example: https://paper.test.broker/place-order
      request:
        type: object
        x-nullable: true
      response:
        type: object
        x-nullable: true
      isError:
        type: boolean
      statusCode:
        type: integer
      duration:
        type: integer
        description: Call duration in ms
      orderId:
        type: string
        x-nullable: true
      brokerOrderId:
        type: string
        x-nullable: true
      userId:
        type: string
        x-nullable: true
      createdAt:
        type: string
        format: date-time
...
  Error:
    type: object
    title: Error
    description: Standard error envelope returned by the central error handler.
    properties:
      error:
        type: object
        properties:
          code:
            type: string
            example: UNAUTHORIZED
          message:
            type: string
            example: Invalid credentials
          requestId:
            type: string
          details:
            type: object
            x-nullable: true
          stack:
            type: string
            description: Present only outside production.